Break up OOP utils to avoid circular dependencies
