+++ lib/eventTarget.js
@@ -3,7 +3,8 @@ var Lang = require('./utilities/lang');
 /**
  * A simple implementation of the observer pattern.
  * Mix me into your class with augment.
- * @method EventTarget
+ * @class EventTarget
+ * @constructor
  */
 var EventTarget = function EventTarget() {
   this.__topics = {};
@@ -28,17 +29,34 @@ Event.prototype = {
 };
 
 EventTarget.prototype = {
+  /**
+   * Fired before the default function has occurred.
+   * Call e.preventDefault() on the event and the action we be prevented.
+   * @method on
+   * @param  {String} topic event to listen to
+   * @param  {Function} callback
+   * @param  {Object} [context]
+   * @param  [...more args...] can be provided and will be passed to the callback}
+   */
   on: function(topic, callback, context /*adtl args*/) {
     return this._registerEvent('on', topic, callback, context, Array.prototype.slice.call(arguments, 3));
   },
 
+  /**
+   * Fired after the default function has occurred.
+   * @method after
+   * @param  {String} topic event to listen to
+   * @param  {Function} callback
+   * @param  {Object} [context]
+   * @param  [...more args...] can be provided and will be passed to the callback}
+   */
   after: function(topic, callback, context /*adtl args*/) {
     return this._registerEvent('after', topic, callback, context, Array.prototype.slice.call(arguments, 3));
   },
 
-  once: function(topic, callback, context /*adtl args*/) {
-    return this._registerEvent('once', topic, callback, context, Array.prototype.slice.call(arguments, 3));
-  },
+  // once: function(topic, callback, context /*adtl args*/) {
+  //   return this._registerEvent('once', topic, callback, context, Array.prototype.slice.call(arguments, 3));
+  // },
 
   _registerEvent: function(type, topic, callback, context, adtlArgs) {
     var listener = {
@@ -57,6 +75,14 @@ EventTarget.prototype = {
 
   },
 
+  /**
+   * Fire an event on this object.
+   * @method fire
+   * @param  {String} topic
+   * @param  {Function} defaultFn a function that will be 
+   *   executed unless the default function was cancelled by a listener
+   * @param  {Object} adtlProperties will be added to the event object
+   */
   fire: function(topic, defaultFn, adtlProperties) {
     var eventFacade = new Event(topic, this);
     if (Lang.isObject(adtlProperties)) {
@@ -98,7 +124,6 @@ EventTarget.prototype = {
 
 
   detachAll: function() {
-    delete this.__topics;
     this.__topics = {};
   }
