+++ lib/utilities/lang.js
@@ -101,7 +101,23 @@ var Lang = {
 
   isBoolean: function(o) {
     return typeof o === 'boolean';
-  }
+  },
+
+  /**
+   * Is it an object? Note that arrays are objects.
+   * @method isObject
+   * @param  o for testing 
+   * @param  {Boolean} [failfn] fail on functions
+   * @return {Boolean}
+   */
+  isObject: function(o, failfn) {
+    var t = typeof o;
+    return (o && (t === 'object' ||
+      (!failfn && Lang.isFunction(o)))) || false;
+  },
+
+  isArray: Array.isArray
+
 
 };
